Create simple box device for memory input/output (#70)
The envisioned use of this class is where you wish to simulate a network interface but the internal hardware logic is either not needed or very simple. A custom adapter can be made for the network interface and can simply read and write to an IoBox. I wondered if a lot of people would write devices like this and if it would make sense for tickit to provide a simple helper device.

Any thoughts welcome.

---------

Co-authored-by: Garry O'Donnell <garry.o'donnell@diamond.ac.uk>