Simulation Wrapper (#142)
This change shifts the simulation builder logic from the cli to a dedicated class, TickitSimulation. It adds a builder function to allow simulations to be constructed in a declarative fashion, allowing for further customisation in the future and better testing now. It is a step towards the larger changes required for simulating zebra devices and uniting master and nested simulation logic.