Add integration tests (#153)
Check the master and slave scheduler in context with other components

Two test cases are set up; one for a regular simulation with a master
scheduler, sink and source. Another for a simulation with a system
simulation component. A fixture sets up the simulation and yields it
before the tests, tearing it down afterwards by throwing an exception
and device interrupt.

---------

Co-authored-by: Abigail Emery <abigail.emery@diamond.ac.uk>
Co-authored-by: DiamondJoseph <53935796+DiamondJoseph@users.noreply.github.com>