Catch HTTP adapter run_forever failures in test fixture (#165)
Currently on master, running pytest if port 8080 on localhost is bound
freezes the output. The terminal simply hangs. I have investigated the
issue and fixed this by changing the fixture where this is happening so
that it keeps track of the task that calls `.run_forever` on the adaptor
as well as `.wait_until_ready` and waits until one of those completes.

wait_until_ready will only ever be set if run_forever gets far enough to
set it; if, for example, the server cannot be instantiated, this fails
silently, and wait_until_ready therefore waits forever.