Adapter rework for Zebra (#174)
This changes the way adapters work within tickit. This
attempts unifies the adapters in a composed way by seperating out the IO
and the device specific handling logic. You now have an 
`AdapterContainer` which has both an `adapter`, the device specific part that handles
the message,  and the `io` which does the actual io logic.

The `AdapterContainer` acts as the previous adapters did and contains
the `run_forever` method. Now though, this calls a `setup` method in the
`io` in which the `io` takes the specific `adapter` that has been
provided. The `adapter` and `io` are failry coupled given the io needs a
specific adapter type to work, eg. `class
HttpIo(AdapterIo[HttpAdapter]): `. However hopefully this change
provides a clearer divide in developer and user implemented code and
simplifies the life of the users.

As before users would need to write specific adapters for their devices
but this should simplifiy their lives a bit as all they now need is the
device and device specific methods. 

Most importantly, and the motivation for this change, is there is no
longer a limitation on what an adapter may apply to. This allows us to
write an adapter which inspect the components in a system simulation
There is an example which uses a tcpio and commandadapter.