Remove broken links (#182)
As described in #181 , the docs build has broken with the latest release
of setuptools. As this is only a docs break, and is creating a blocker
for more urgent work I suggest we remove them for now to allow CI to
pass, then dedicate the time to fix it correctly after the changed for
Zebra are implemented.